Remove PoxPaginate::Root requires.
This flies in the face of rails require process since it will yield paths
that in some cases are relative and some that are fully qualified. Most notable
is the version file which is evaluated as part of bundler and then when the
rails app boots, you will get const warning for already defined MAJOR, MINOR,
etc. Just do the right thing and require paths that are in the load path and
solved higher up.